Fix 0 card calculation.
When there is 0 cards, we just return no values. Java by default gives an integer value 0 if it is not instantiated, but I needed to explicitly set the score to 0 so when I edit a game to have both players be 0, it will recalculate the score and recalculate the winning players and reset the winner text, icon, etc.