Save data complete.
When the application is exited (through the back button on main activity or home button), it will save the data.

In the instance that the app has not been fully exited (not cleared from RAM/recent apps), gameInfoCardModels arraylist is cleared to prevent duplicates when data is reloaded (app does not create a new instance so data is duplicated if not cleared).